+++ botReborn.cabal
@@ -1,6 +1,6 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.35.1.
+-- This file has been generated from package.yaml by hpack version 0.36.0.
 --
 -- see: https://github.com/sol/hpack
 
@@ -53,7 +53,7 @@ library
       OverloadedStrings
       DeriveGeneric
       LambdaCase
-  ghc-options: -Wall -Werror -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints
+  ghc-options: -Wall
   build-depends:
       aeson
     , base >=4.7 && <5
@@ -74,7 +74,7 @@ executable botReborn-exe
       OverloadedStrings
       DeriveGeneric
       LambdaCase
-  ghc-options: -Wall -Werror -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
+  ghc-options: -Wall -threaded -rtsopts -with-rtsopts=-N
   build-depends:
       aeson
     , base >=4.7 && <5
@@ -88,6 +88,10 @@ test-suite botReborn-test
   type: exitcode-stdio-1.0
   main-is: Spec.hs
   other-modules:
+      Handlers.BaseSpec
+      Handlers.BotSpec
+      Handlers.DispatcherSpec
+      Handlers.LoggerSpec
       Paths_botReborn
   hs-source-dirs:
       test
@@ -95,7 +99,7 @@ test-suite botReborn-test
       OverloadedStrings
       DeriveGeneric
       LambdaCase
-  ghc-options: -Wall -Werror -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
+  ghc-options: -Wall -threaded -rtsopts -with-rtsopts=-N
   build-depends:
       QuickCheck
     , base >=4.7 && <5
